requirements.txt
flask
flask-sqlalchemy
flask-login
flask-bcrypt
mysql-connector-python
tensorflow
keras
numpy
pandas
opencv-python
pillow
matplotlib
fpdf
scikit-learn
flask-migrate
python-dotenv
flask-wtf
flask-talisman
google-generativeai
email-validator
